# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MedCenter Back",
    "description": "API for MedCenter",
    "version": "1.0.0"
  },
  "host": "localhost:8088",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/info": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Получение информации о пользователе",
        "description": "Получение полной информации о пользователе",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Не удалось получить информацию по пользователю",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/update-password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Обновление пароля",
        "description": "Обновление пароля пользователя",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Получение нового пароля",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "new_password": {
                  "type": "string",
                  "example": "Новый пароль"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пароль успешно обновлен",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось обновить пароль",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/profile": {
      "patch": {
        "tags": ["Users"],
        "summary": "Обновление персональных данных",
        "description": "Обновление имени, email, телефона, номера полиса текущего пользователя (пациента). Все поля опциональны.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/UpdateProfile" }
          }
        ],
        "responses": {
          "200": { "description": "Данные обновлены", "schema": { "$ref": "#/definitions/User" } },
          "400": { "description": "Ошибка валидации или email занят", "schema": { "$ref": "#/definitions/Message" } },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      }
    },
    "/me/history": {
      "get": {
        "tags": ["Users"],
        "summary": "Моя медицинская история",
        "description": "Получение всей медицинской истории текущего пользователя (по id пользователя). Объединяет записи по всем картам пациента. Не требует id карты. Требуется авторизация.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Список записей истории",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/MedicalHistoryRecord" }
            }
          },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      }
    },
    "/me/appointments": {
      "get": {
        "tags": ["Users"],
        "summary": "Мои записи (приёмы к врачам)",
        "description": "Список всех приёмов текущего пользователя. По всем картам (докторам). Сортировка: дата по убыванию, время. Требуется авторизация.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Список приёмов",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/AppointmentRecord" }
            }
          },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Создание приёма к врачу",
        "description": "Запись на приём к врачу. Если карты пациента у этого доктора ещё нет — она создаётся автоматически. Статус новой записи — «Ожидает подтверждения» (pending). Требуется авторизация.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CreateAppointment" }
          }
        ],
        "responses": {
          "201": {
            "description": "Приём создан",
            "schema": { "$ref": "#/definitions/AppointmentRecord" }
          },
          "400": {
            "description": "Ошибка валидации или нет карты у этого доктора",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      }
    },
    "/doctor": {
      "post": {
        "tags": [
          "Doctors"
        ],
        "summary": "Создание нового доктора",
        "description": "Создание нового пользователя с типом доктор (только для администраторов)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Данные нового доктора: email, password, fullName — обязательны; остальные поля опциональны.",
            "schema": {
              "$ref": "#/definitions/CreateDoctor"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Доктор успешно создан",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Ошибка валидации или email уже занят",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "401": {
            "description": "Требуются права администратора",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/doctor/{id}": {
      "patch": {
        "tags": [
          "Doctors"
        ],
        "summary": "Обновление информации о докторе",
        "description": "Обновление данных доктора по id (только для администраторов)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID доктора"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Поля для обновления. Передавайте только те поля, которые нужно изменить; все опциональны.",
            "schema": {
              "$ref": "#/definitions/UpdateDoctor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Данные доктора обновлены",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Ошибка валидации",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "401": {
            "description": "Требуются права администратора",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "404": {
            "description": "Доктор не найден",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      },
      "delete": {
        "tags": ["Doctors"],
        "summary": "Удаление доктора",
        "description": "Удаление доктора по id. Пациенты открепляются от доктора. Только для администраторов.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID доктора"
          }
        ],
        "responses": {
          "200": {
            "description": "Доктор удалён",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "400": {
            "description": "Некорректный id",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "401": {
            "description": "Требуются права администратора",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "404": {
            "description": "Доктор не найден",
            "schema": { "$ref": "#/definitions/Message" }
          }
        }
      }
    },
    "/doctor/{id}/avatar": {
      "patch": {
        "tags": ["Doctors"],
        "summary": "Загрузка аватара доктора",
        "description": "Загрузка/обновление аватара доктора. Доступ: администратор или сам доктор (по id). Формат: multipart/form-data, поле «image» (jpeg, png, jpg).",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID доктора"
          },
          {
            "name": "image",
            "in": "formData",
            "required": true,
            "type": "file",
            "description": "Файл изображения (jpeg, png, jpg)"
          }
        ],
        "responses": {
          "200": { "description": "Аватар обновлён", "schema": { "$ref": "#/definitions/User" } },
          "400": { "description": "Файл не передан или некорректный", "schema": { "$ref": "#/definitions/Message" } },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } },
          "403": { "description": "Нет прав на обновление аватара", "schema": { "$ref": "#/definitions/Message" } },
          "404": { "description": "Доктор не найден", "schema": { "$ref": "#/definitions/Message" } }
        }
      }
    },
    "/doctor/{id}/available-slots": {
      "get": {
        "tags": ["Doctors"],
        "summary": "Доступные даты и слоты доктора",
        "description": "Список всех доступных дат в диапазоне [from, to], когда доктор принимает (dayWork), с вложенными свободными слотами по 30 минут (scheduleWork). Учитываются уже занятые записи (pending/confirmed). Если from и to не переданы — период от сегодня до +1 месяц.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID доктора"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date",
            "description": "Начало периода YYYY-MM-DD"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date",
            "description": "Конец периода YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Доступные даты и часы приёма",
            "schema": {
              "type": "object",
              "properties": {
                "available": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": { "type": "string", "example": "2024-01-24" },
                      "slots": {
                        "type": "array",
                        "items": { "type": "string", "example": "12:00" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный id или from позже to",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "404": {
            "description": "Доктор не найден",
            "schema": { "$ref": "#/definitions/Message" }
          }
        }
      }
    },
    "/doctor/{id}/pacients": {
      "get": {
        "tags": ["Doctors"],
        "summary": "Получение пациентов доктора",
        "description": "Список всех пациентов, закреплённых за доктором. Только для администраторов.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID доктора"
          }
        ],
        "responses": {
          "200": {
            "description": "Список пациентов",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/PacientItem" }
            }
          },
          "400": {
            "description": "Некорректный id",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "401": {
            "description": "Требуются права администратора",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "404": {
            "description": "Доктор не найден",
            "schema": { "$ref": "#/definitions/Message" }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Message": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Сообщение"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        },
        "email": {
          "type": "string",
          "example": "test@mail.ru"
        },
        "fullName": {
          "type": "string",
          "example": "user"
        },
        "phone": {
          "type": "string",
          "example": "+7 999 123-45-67"
        },
        "policyNumber": {
          "type": "string",
          "description": "Номер полиса (мед. полис)",
          "example": "1234567890123456"
        },
        "isAdmin": {
          "type": "boolean",
          "example": false
        },
        "createdAt": {
          "type": "string",
          "example": "24-12-2025"
        },
        "updatedAt": {
          "type": "string",
          "example": "24-12-2025"
        },
        "userType": {
          "default": "DOCTOR",
          "type": "string",
          "enum": [
            "DOCTOR",
            "PACIENT",
            "REGISTER",
            "MANAGER"
          ]
        }
      }
    },
...
    "CreateAppointment": {
      "type": "object",
      "required": ["doctorId", "dateVisit", "time"],
      "properties": {
        "doctorId": { "type": "integer", "description": "ID доктора" },
        "dateVisit": { "type": "string", "format": "date", "description": "Дата приёма (YYYY-MM-DD)", "example": "2024-02-20" },
        "time": { "type": "string", "description": "Время приёма (ЧЧ:ММ)", "example": "10:00" },
        "roomNumber": { "type": "string", "maxLength": 50, "description": "Номер кабинета (опционально)", "example": "205" }
      }
    },
    "UpdateProfile": {
      "type": "object",
      "description": "Персональные данные для обновления (все поля опциональны)",
      "properties": {
        "fullName": { "type": "string", "maxLength": 255, "example": "Иванов Иван" },
        "email": { "type": "string", "format": "email", "example": "user@mail.ru" },
        "phone": { "type": "string", "maxLength": 20, "example": "+7 999 123-45-67" },
        "policyNumber": { "type": "string", "maxLength": 32, "description": "Номер полиса", "example": "1234567890123456" }
      }
    },
...
  }
}